backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email-validation]==2.9.0
scikit-learn
pandas
numpy
python-dotenv==1.0.1
alembic==1.13.2

firebase-admin==6.5.0
google-cloud-bigquery==3.25.0
